refactoring + jsdoc updates

+++ packages/cmrjs/src/cmr-utils.js
@@ -240,16 +240,16 @@ function granulesToCmrFileObjects(granules, filterFunc = isCMRFile) {
 }
 
 /**
- * Run a CMR operation with retry logic for launchpad failures. If the operation fails with a
+ * Runs a CMR operation with retry logic for launchpad failures. If the operation fails with a
  * 401, refresh the Launchpad token and retry.
  *
- * @param {CMR} cmrClient - the CMR client whose token should be refreshed on 401
+ * @param {CMR} cmrClient - the CMR client whose launchpad token should be refreshed on 401
  * @param {() => Promise} operation - the CMR function with args to execute
  * @param {Object} [options]
  * @param {number} [options.retries=10] - number of retry attempts
  * @returns {Promise} - result of CMR function call
  */
-async function withCmrTokenRefreshRetry(cmrClient, operation, { retries = 10 } = {}) {
+async function withCmrLaunchpadTokenRefreshRetry(cmrClient, operation, { retries = 10 } = {}) {
   return await pRetry(
     async () => {
       try {
@@ -278,13 +278,18 @@ async function withCmrTokenRefreshRetry(cmrClient, operation, { retries = 10 } =
  * @returns {Promise<Object>} CMR's success response which includes the concept-id
  */
 async function publishECHO10XML2CMR(cmrFile, cmrClient, revisionId) {
+  let res;
   const builder = new xml2js.Builder();
   const xml = builder.buildObject(cmrFile.metadataObject);
 
-  const res = await withCmrTokenRefreshRetry(
-    cmrClient,
-    () => cmrClient.ingestGranule(xml, revisionId)
-  );
+  if (cmrClient.oauthProvider === 'launchpad') {
+    res = await withCmrLaunchpadTokenRefreshRetry(
+      cmrClient,
+      () => cmrClient.ingestGranule(xml, revisionId)
+    );
+  } else {
+    res = await cmrClient.ingestGranule(xml, revisionId);
+  }
 
   const conceptId = res.result['concept-id'];
   let resultLog = `Published ${cmrFile.granuleId} to the CMR. conceptId: ${conceptId}`;
@@ -313,12 +318,17 @@ async function publishECHO10XML2CMR(cmrFile, cmrClient, revisionId) {
  * @returns {Promise<Object>} CMR's success response which includes the concept-id
  */
 async function publishUMMGJSON2CMR(cmrFile, cmrClient, revisionId) {
+  let res;
   const granuleId = cmrFile.metadataObject.GranuleUR;
 
-  const res = await withCmrTokenRefreshRetry(
-    cmrClient,
-    () => cmrClient.ingestUMMGranule(cmrFile.metadataObject, revisionId)
-  );
+  if (cmrClient.oauthProvider === 'launchpad') {
+    res = await withCmrLaunchpadTokenRefreshRetry(
+      cmrClient,
+      () => cmrClient.ingestUMMGranule(cmrFile.metadataObject, revisionId)
+    );
+  } else {
+    res = await cmrClient.ingestUMMGranule(cmrFile.metadataObject, revisionId);
+  }
 
   const conceptId = res['concept-id'];
 
@@ -377,10 +387,13 @@ function publish2CMR(cmrPublishObject, creds, cmrRevisionId) {
  */
 async function removeFromCMR(granuleUR, creds) {
   const cmrClient = new CMR(creds);
-  return await withCmrTokenRefreshRetry(
-    cmrClient,
-    () => cmrClient.deleteGranule(granuleUR)
-  );
+  if (cmrClient.oauthProvider === 'launchpad') {
+    return await withCmrLaunchpadTokenRefreshRetry(
+      cmrClient,
+      () => cmrClient.deleteGranule(granuleUR)
+    );
+  }
+  return await cmrClient.deleteGranule(granuleUR);
 }
 
 /**
@@ -993,7 +1006,8 @@ async function getCmrSettings(cmrConfig = {}) {
   };
 
   if (oauthProvider === 'launchpad') {
-    // checking for lock file and waiting for its release if it exists
+    // checking for lock file and waiting for its release if it exists in case a new token
+    // is being created by another CMR process due to launchpad auth failures
     await launchpad.waitForLockFileRelease();
 
     const launchpadPassphraseSecretName = cmrConfig.passphraseSecretName
@@ -1009,9 +1023,7 @@ async function getCmrSettings(cmrConfig = {}) {
     };
 
     log.debug('cmrjs.getCreds getLaunchpadToken');
-
     const token = await launchpad.getLaunchpadToken(config);
-
     return {
       ...cmrCredentials,
       ...config,
@@ -1406,19 +1418,36 @@ function buildCMRQuery(results) {
  * containing the found collections
  */
 async function getCollectionsByShortNameAndVersion(results) {
+  let response;
   const query = buildCMRQuery(results);
   const cmrClient = new CMR(await getCmrSettings());
-  const response = await withCmrTokenRefreshRetry(cmrClient, async () => {
+  if (cmrClient.oauthProvider === 'launchpad') {
+    response = await withCmrLaunchpadTokenRefreshRetry(cmrClient, async () => {
+      const headers = cmrClient.getReadHeaders({ token: await cmrClient.getToken() });
+      return got.post(
+        `${getSearchUrl()}collections.json`,
+        {
+          json: query,
+          responseType: 'json',
+          headers: { Accept: 'application/json', ...headers },
+        }
+      );
+    });
+  } else {
     const headers = cmrClient.getReadHeaders({ token: await cmrClient.getToken() });
-    return got.post(
+
+    response = await got.post(
       `${getSearchUrl()}collections.json`,
       {
         json: query,
         responseType: 'json',
-        headers: { Accept: 'application/json', ...headers },
+        headers: {
+          Accept: 'application/json',
+          ...headers,
+        },
       }
     );
-  });
+  }
   return response.body;
 }
